% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARemoval.R
\name{clean_data}
\alias{clean_data}
\title{Function reads in data set, omits any possible NA values, and returns that.}
\usage{
clean_data(filename)
}
\arguments{
\item{filename}{name of file}
}
\value{
The NA values omitted from filename
}
\description{
Function reads in data set, omits any possible NA values, and returns that.
}
\examples{
clean_data(BatData)
}
