% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistogramPlot.R
\name{histo_plot}
\alias{histo_plot}
\title{Function reads in data set, creates a histogram, and returns that.}
\usage{
histo_plot(dataframe)
}
\arguments{
\item{filename}{name of file}
}
\value{
Histogram of selected data set comparing heart rate against energy being expended.
}
\description{
Function reads in data set, creates a histogram, and returns that.
}
\examples{

histo_plot(BatData)

}
