% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerPlot.R
\name{Joules}
\alias{Joules}
\title{Energy Expenditure and Heart Rate Power Curve}
\usage{
Joules(dataframe)
}
\arguments{
\item{dataframe}{name of dataframe}
}
\value{
Power curve from dataframe.
}
\description{
Function that generates a Power Curve visualization derived from energy expenditure and heart rate metrics from dataframe.
}
\examples{
Joules(accdata)
}
