% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappedPoints.R
\name{Mapping}
\alias{Mapping}
\title{Function to map data points of latitude and longitude from coordinates obtained from collected GPS data.}
\usage{
Mapping(dataframe)
}
\arguments{
\item{dataframe}{name of dataframe}
}
\value{
Mapped data points of collected GPS coordinates of bats in Panama.
}
\description{
Function to map data points of latitude and longitude from coordinates obtained from collected GPS data.
}
\examples{
Mapping(batcoord)
}
