% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardCalcs.R
\name{sdfh}
\alias{sdfh}
\title{Function calculates the standard deviation of a basic spatial data set and returns that.}
\usage{
sdfh(dataframe)
}
\arguments{
\item{filename}{name of file}
}
\value{
standard deviation
}
\description{
Function calculates the standard deviation of a basic spatial data set and returns that.
}
\examples{
sdfh(BatData)
}
