% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerPlot.R
\name{Joules}
\alias{Joules}
\title{Function to calculate energy exerted compared to heart rate.}
\usage{
Joules(dataframe)
}
\arguments{
\item{filename}{name of file}
}
\value{
Line graph of selected data set comparing heart rate against energy being expended.
}
\description{
Function to calculate energy exerted compared to heart rate.
}
\examples{
Joules(accdata)
}
