% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingPoints.R
\name{Mapping}
\alias{Mapping}
\title{Mapping spatial data points}
\usage{
Mapping(dataframe)
}
\arguments{
\item{dataframe}{name of dataframe}
}
\value{
Map generated from spatial data points in dataframe
}
\description{
Function maps spatial data points on interactive map.
}
\examples{
Mapping(batcoord)
}
