% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardCalcs.R
\name{sdfh}
\alias{sdfh}
\title{Standard Deviation of Heart Rates}
\usage{
sdfh(dataframe)
}
\arguments{
\item{dataframe}{name of dataframe}
}
\value{
Standard deviation will be calculated from dataframe.
}
\description{
Function calculates standard deviation of dataframe, and returns that.
}
\examples{
sdfh(HeartRate)
}
